--- src/layers/ZweitausbildungRoutesHighlightLayer2/index.js
@@ -1,1 +1,0 @@
-export { default } from './ZweitausbildungRoutesHighlightLayer2';
