--- src/layers/ZweitausbildungRoutesLayer2/index.js
@@ -1,1 +1,0 @@
-export { default } from './ZweitausbildungRoutesLayer2';
